# Map Inf path lengths to zero