Issue opened in #116.

Closes #116