@Datseris 2021-05-21
Thank you :) @gerrygralton Could you please provide a unit test for this scenario? You should put the test in one of the files in the `test` folder. Reading any file will guide you on how test work.

@heliosdrm  or  @hkraemer would you kindly review and approve the PR? I do not have the mathematical background :)