@heliosdrm 2021-05-21
Thanks! If you have a specific example of a network with disconected nodes that you want to test, either defined as a `LightGraph` or by its adjacency matrix, you may add it in the last block of `test/smallmatrix.jl`, where there are other examples that you can follow.