@gerrygralton 2021-05-24
Hi, 
Thanks for having a look at this! In 127dad8a8de6679d8e09f352e7dd5dab2e1aec15 I've changed the formula slightly to remove the LightGraphs.closeness_centrality dependency because that formulation really doesn't like disconnected graphs.
It also has the advantage of being faster. Using the 10x10 example from test/small_matrix.jl:

New:
    
    BenchmarkTools.Trial: 
      memory estimate:  2.67 KiB
      allocs estimate:  5
      --------------
      minimum time:     929.786 ns (0.00% GC)
      median time:      957.500 ns (0.00% GC)
      mean time:        1.104 μs (5.17% GC)
      maximum time:     101.671 μs (98.23% GC)
      --------------
      samples:          10000
      evals/sample:     28
  
Old:

    BenchmarkTools.Trial: 
      memory estimate:  20.08 KiB
      allocs estimate:  165
      --------------
      minimum time:     11.829 μs (0.00% GC)
      median time:      12.584 μs (0.00% GC)
      mean time:        15.399 μs (8.92% GC)
      maximum time:     3.706 ms (98.96% GC)
      --------------
      samples:          10000
      evals/sample:     1

I have also added some tests in 989bd14dd8029873409c986c24f81662a350a91b.