@gerrygralton 2021-05-24
> Thanks, can you please adjust Project.toml version and CHANGELOG.md accordingly?

Uh, there is no CHANGELOG.md...