@Datseris 2021-05-24
Woooooooooooops sorry about that, let's start one then. Just provide the note that changelog starts being kept from current version.