@Datseris 2021-05-24
Ah, sorry for the confusion @gerrygralton . Yeap, it is NEWS.md. I got confused, in some other packages it is called CHANGELOG.md...